impl(py27): import depending on interpreter version
